Extracted thumbnail logic from core.py into a new thumbnail module.
